+++ src/app/app.js
@@ -35,7 +35,7 @@ function preformattedDetails(response) {
 }
 
 
-function checkResponse(response, theme) {
+function checkResponse(response, theme, debug) {
   if (!response.ok) {
     return Promise.resolve(
       () => (
@@ -50,7 +50,15 @@ function checkResponse(response, theme) {
   return response.text().then((text) => {
     try {
       const config = JSON.parse(text);
-      return Promise.resolve(() => (<Vitessce config={config} theme={theme} />));
+      return Promise.resolve(() => (
+        <Vitessce
+          config={config}
+          theme={theme}
+          // eslint-disable-next-line no-console
+          onConfigChange={debug ? console.log : undefined}
+          validateOnConfigChange={debug}
+        />
+      ));
     } catch (e) {
       return Promise.resolve(() => (
         <Warning
@@ -73,7 +81,18 @@ function validateTheme(theme) {
   return (['light', 'dark'].includes(theme) ? theme : 'dark');
 }
 
-export function createApp(rowHeight = null) {
+/**
+ * Convenience function for creating the minimal Vitessce demo and demo listing
+ * components based on the current URL parameters.
+ * @param {object} params
+ * @param {number|null} params.rowHeight The row height to pass to the Vitessce grid.
+ * Optional. By default, null.
+ * @param {boolean} showBetaHeader Should the header which links to the beta documentation
+ * website be rendered? Optional. By default, false.
+ * @returns A component, either <Welcome/> or <Vitessce/> depending on the URL params.
+ */
+export function createApp(params) {
+  const { rowHeight = null, showBetaHeader = false } = params;
   const urlParams = new URLSearchParams(window.location.search);
   const datasetId = urlParams.get('dataset');
   const debug = urlParams.get('debug') === 'true';
@@ -96,7 +115,7 @@ export function createApp(rowHeight = null) {
   }
   if (datasetUrl) {
     const responsePromise = fetch(datasetUrl)
-      .then(response => checkResponse(response, theme))
+      .then(response => checkResponse(response, theme, debug))
       .catch(error => Promise.resolve(
         <Warning
           title="Error fetching"
@@ -109,5 +128,5 @@ export function createApp(rowHeight = null) {
     );
   }
   const configs = listConfigs(showAll);
-  return (<Welcome configs={configs} theme={theme} />);
+  return (<Welcome configs={configs} theme={theme} showBetaHeader={showBetaHeader} />);
 }