+++ src/components/layer-controller/RasterLayerController.js
@@ -46,7 +46,12 @@ export default function RasterLayerController(props) {
     handleLayerRemove, handleLayerChange,
   } = props;
 
-  const { colormap, opacity, channels } = layer;
+  const {
+    colormap,
+    opacity,
+    channels,
+    transparentColor,
+  } = layer;
   const firstSelection = channels[0]?.selection || {};
 
   const { dimensions } = loader;
@@ -67,6 +72,9 @@ export default function RasterLayerController(props) {
 
   function setChannels(v) {
     handleLayerChange({ ...layer, channels: v });
+  }
+  function setTransparentColor(v) {
+    handleLayerChange({ ...layer, transparentColor: v });
   }
 
   function setChannelsAndDomainType(newChannels, newDomainType) {
@@ -252,6 +260,7 @@ export default function RasterLayerController(props) {
             dimensions={dimensions}
             opacity={opacity}
             colormap={colormap}
+            transparentColor={transparentColor}
             domainType={domainType}
             // Only allow for global dimension controllers that
             // exist in the `dimensions` part of the loader.
@@ -266,6 +275,7 @@ export default function RasterLayerController(props) {
             handleGlobalChannelsSelectionChange={
               handleGlobalChannelsSelectionChange
             }
+            handleTransparentColorChange={setTransparentColor}
             isRgb={loader.isRgb}
             handleDomainChange={handleDomainChange}
           />