+++ src/components/spatial/SpatialSubscriber.js
@@ -4,8 +4,13 @@ import { capitalize } from '../../utils';
 import { useDeckCanvasSize, useReady, useUrls } from '../hooks';
 import { setCellSelection, mergeCellSets } from '../utils';
 import {
-  useCellsData, useCellSetsData, useExpressionMatrixData,
-  useMoleculesData, useNeighborhoodsData, useRasterData,
+  useCellsData,
+  useCellSetsData,
+  useGeneSelection,
+  useMoleculesData,
+  useNeighborhoodsData,
+  useRasterData,
+  useExpressionAttrs,
 } from '../data-hooks';
 import { getCellColors } from '../interpolate-colors';
 import Spatial from './Spatial';
@@ -141,8 +146,11 @@ export default function SpatialSubscriber(props) {
     { setCellSetSelection, setCellSetColor },
     { initializeCellSetSelection, initializeCellSetColor },
   );
-  const [expressionMatrix] = useExpressionMatrixData(
-    loaders, dataset, setItemIsReady, addUrl, false,
+  const [expressionData] = useGeneSelection(
+    loaders, dataset, setItemIsReady, false, geneSelection,
+  );
+  const [attrs] = useExpressionAttrs(
+    loaders, dataset, setItemIsReady, addUrl, true,
   );
   // eslint-disable-next-line no-unused-vars
   const [raster, imageLayerLoaders] = useRasterData(
@@ -176,13 +184,14 @@ export default function SpatialSubscriber(props) {
 
   const cellColors = useMemo(() => getCellColors({
     cellColorEncoding,
-    expressionMatrix,
+    expressionData: expressionData && expressionData[0],
     geneSelection,
     cellSets: mergedCellSets,
     cellSetSelection,
     cellSetColor,
+    expressionDataAttrs: attrs,
   }), [cellColorEncoding, geneSelection, mergedCellSets,
-    cellSetColor, cellSetSelection, expressionMatrix]);
+    cellSetColor, cellSetSelection, expressionData, attrs]);
 
   const cellSelection = useMemo(() => Array.from(cellColors.keys()), [cellColors]);